Removed security from gateway.
