--- .gitmodules
@@ -1,3 +1,0 @@
-[submodule "ORM-Library"]
-	path = ORM-Library
-	url = https://github.com/QuantumQuackDoctor/ORM-Library.git
