--- SSGateway/src/main/java/com/smoothstack/SSGateway/filter/RouterValidator.java
@@ -1,20 +1,0 @@
-package com.smoothstack.SSGateway.filter;
-
-import org.springframework.http.server.reactive.ServerHttpRequest;
-import org.springframework.stereotype.Component;
-
-import java.util.Arrays;
-import java.util.List;
-import java.util.function.Predicate;
-
-@Component
-public class RouterValidator {
-
-    public static final List<String> openApiEndpoints = Arrays.asList(
-            "/login",
-            "/register"
-    );
-
-    public Predicate<ServerHttpRequest> isSecured = request ->
-            openApiEndpoints.stream().noneMatch(uri -> request.getURI().getPath().contains(uri));
-}
