--- SSGateway/src/test/java/com/smoothstack/SSGateway/SsGatewayApplicationTests.java
@@ -1,13 +1,0 @@
-package com.smoothstack.SSGateway;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class SsGatewayApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
