+++ src/main/java/com/smoothstack/SSGateway/SsGatewayApplication.java
@@ -1,0 +1,23 @@
+package com.smoothstack.SSGateway;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.gateway.route.RouteLocator;
+import org.springframework.cloud.gateway.route.builder.RouteLocatorBuilder;
+import org.springframework.context.annotation.Bean;
+
+@SpringBootApplication
+public class SsGatewayApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(SsGatewayApplication.class, args);
+    }
+
+    @Bean
+    public RouteLocator gatewayRoutes(RouteLocatorBuilder builder) {
+        return builder.routes()
+                .route(r -> r.path("/user", "/users", "/login", "/register")
+                        .uri("http://localhost:8080"))
+                .build();
+    }
+}
