+++ src/main/resources/application.properties
@@ -9,6 +9,3 @@ eureka.client.refresh.enable=true
 spring.application.name=gateway
 
 server.port=3000
-
-#this is only used in testing
-jwt.secret=testSecret
